#258: cleanse docs of code references; minimal comments; split UploadStorageBackend
Principle 1 — docs describe concepts, not code:
- docs/glossary.md: removed UploadHandle entry; cleansed Atomic destination reservation
  of all class/API/expect-actual names; cleansed Abort of UploadHandle reference
- docs/engineering/testing.md: removed code names (TrustedDeviceStore, File,
  DataStore<Preferences>) from jvmTest description
- docs/engineering/file-transfer-wire.md: removed all class names from Storage seam
  section and Layer 2 heading; described mechanism conceptually

Principle 2 — minimal comments, no code-restatement:
- AppleUploadStorageBackendTest.kt: deleted class-level race-branch deflection block
  and two inline restatement comments
- JvmUploadStorageBackendTest.kt: deleted two inline restatement comments
- UploadStorageBackend.kt: deleted KDoc that mentioned Android MediaStore
- AtomicFileReserver.kt: trimmed KDoc to postcondition only

Principle 3 — commonMain single-responsibility files:
- Split UploadStorageBackend.kt into UploadStorageBackend.kt (interface) and
  FileUploadStorage.kt (FileUploadStorage class + mkdirsTracked helper)
