+++ composeApp/src/appleTest/kotlin/com/tubetoast/tether/network/AppleUploadStorageBackendTest.kt
@@ -11,14 +11,6 @@ import kotlin.test.assertFailsWith
 import kotlin.test.assertFalse
 import kotlin.test.assertTrue
 
-// The concurrent-sibling branch (`createDirectory` fails because a sibling raced
-// between `fileExistsAtPath` and `createDirectory`) is not deterministically
-// reproducible without injecting a seam around the NSFileManager calls. The
-// `mkdirIfAbsent` contract for that branch is covered by the JVM equivalent
-// (where `File.mkdir() == false` is observable) and by the integration race in
-// FileServerConcurrencyTest. Mirroring the test set with the JVM backend keeps
-// the documented contract paths (newly-created, already-exists, externally-created,
-// missing-parent) explicit on both platforms.
 class AppleUploadStorageBackendTest {
     private val tempDirs = mutableListOf<String>()
 
@@ -60,7 +52,6 @@ class AppleUploadStorageBackendTest {
 
     @Test
     fun mkdirIfAbsent_returns_false_when_directory_was_created_externally_before_the_call() {
-        // Covers the fast-path branch: fileExistsAtPath == true before createDirectory is attempted.
         val root = newTempDir()
         val backend = AppleUploadStorageBackend(root)
         val target = "$root/pre-created"
@@ -77,7 +68,6 @@ class AppleUploadStorageBackendTest {
     fun mkdirIfAbsent_throws_when_parent_does_not_exist() {
         val root = newTempDir()
         val backend = AppleUploadStorageBackend(root)
-        // createDirectory fails and the dir still doesn't exist → IOException
         val noParent = "$root/nonexistent-parent/child"
         assertFailsWith<Exception> { backend.mkdirIfAbsent(noParent) }
     }
