+++ docs/glossary.md
@@ -35,10 +35,9 @@ Engineering concepts. The vocabulary [`architect`](../.claude/agents/architect.m
 - **Bonjour** — Apple's implementation of mDNS/DNS-SD service discovery, used by Tether's macOS Desktop publisher via the system `mDNSResponder` daemon and JNA bindings. _Avoid:_ saying "Bonjour" when meaning any mDNS implementation — say "mDNS" instead.
 - **Token** — a named value in `TetherTheme` (color, typography, spacing, shape) read via `TetherTheme.<scale>` from the composition rather than hardcoded. _Avoid:_ constant, raw value, magic number, theme value. (see [ui-style-guide.md](engineering/ui-style-guide.md))
 - **UploadStorage** — the per-platform sink the FileServer talks to for resolving destinations, streaming bytes, enforcing the canonical-realisation check, and aborting on failure. _Avoid:_ storage backend, file sink, persistence layer. (see [file-transfer-wire.md](engineering/file-transfer-wire.md))
-- **UploadHandle** — a value returned by `UploadStorage.resolveDestination` carrying the reserved destination path and the list of parent directories created during the call, used to scope `abort` to only the directories this upload owns.
-- **Atomic destination reservation** — the step inside `UploadStorage.resolveDestination` that picks a deduplicated leaf name and atomically creates a zero-byte placeholder file before any byte of the upload body arrives, so two concurrent uploads with the same `name` cannot resolve to the same path. Implemented per-platform (`File.createNewFile` on JVM, `fopen("wbx")` on Apple) behind the `UploadStorageBackend` interface, with `JvmUploadStorageBackend` / `AppleUploadStorageBackend` per platform. _Avoid:_ destination claim, path lock. (see [file-transfer-wire.md](engineering/file-transfer-wire.md))
+- **Atomic destination reservation** — the step that picks a deduplicated leaf name and atomically creates a zero-byte placeholder file before any byte of the upload body arrives, so two concurrent uploads with the same name cannot resolve to the same path. _Avoid:_ destination claim, path lock. (see [file-transfer-wire.md](engineering/file-transfer-wire.md))
 - **Relative POSIX path** — the file's path inside a transfer, using `/` separators, no leading slash, no `..` segments; for a flat send it is the leaf name. _Avoid:_ filename when nesting matters, relative URL.
 - **Path sanitization** — the two-layer boundary that maps an untrusted `name` parameter to a safe on-disk destination: Layer 1 is the lexical sanitizer in the route handler, Layer 2 is the canonical-realisation check in UploadStorage. (see [file-transfer-wire.md](engineering/file-transfer-wire.md))
-- **Abort** — UploadStorage's failure-path cleanup: deletes the partial destination file and removes only the empty parent directories the upload created (those recorded in the UploadHandle). _Avoid:_ cancel, rollback.
+- **Abort** — UploadStorage's failure-path cleanup: deletes the partial destination file and removes only the empty parent directories this upload created, leaving directories shared with other in-flight uploads alone. _Avoid:_ cancel, rollback.
 - **Downloads root** — the per-device root directory into which incoming files are landed; the security boundary every received file must stay inside. _Avoid:_ destination root, downloads folder when the security property matters.
